requirements.txt
streamlit
pandas
duckdb
plotly
python-dotenv
requests
tenacity
google-analytics-data
google-ads
square
